refactor: Clarify intent of `internal` module
As part of #1410, the name `internal.rs` was discussed and it was
brought up that `parser.rs` would provide clearer intent.
