+++ src/str.rs
@@ -150,7 +150,7 @@ mod test {
     }
   }
 
-  use crate::internal::Needed;
+  use crate::parser::Needed;
 
   fn is_alphabetic(c: char) -> bool {
     (c as u8 >= 0x41 && c as u8 <= 0x5A) || (c as u8 >= 0x61 && c as u8 <= 0x7A)
